+++ models/mmu_cache/lib/mmu_cache.h
@@ -48,16 +48,16 @@
 #ifndef __MMU_CACHE_H__
 #define __MMU_CACHE_H__
 
+#include <greencontrol/config.h>
+#include <amba.h>
 #include <tlm.h>
 #include <tlm_utils/simple_target_socket.h>
 
 #include <math.h>
-#include <greencontrol/config.h>
 
 #include "icio_payload_extension.h"
 #include "dcio_payload_extension.h"
 
-#include <amba.h>
 #include "socrocket.h"
 #include "signalkit.h"
 #include "ahbdevice.h"
@@ -284,10 +284,10 @@ class mmu_cache : public sc_core::sc_module, public mmu_cache_if, public AHBDevi
         
         gs::gs_param_array m_performance_counters;
         /// Total number of successful transactions for execution statistics 
-        gs::gs_param<uint64_t> m_right_transactions;
+        gs::gs_param<unsigned long long> m_right_transactions;
 
 	/// Total number of transactions for execution statistics
-        gs::gs_param<uint64_t> m_total_transactions;
+        gs::gs_param<unsigned long long> m_total_transactions;
 
 	/// power monitoring enabled
 	bool m_pow_mon;
